refactor: wiz뉴스 페이지 컴포넌트 리팩토링
- 컴포넌트 코드 개선
- 추후 작업을 위한 경로 수정
