+++ src/components/Media/NewsPage.tsx
@@ -3,10 +3,11 @@ import { flexColumn, flexRow, flexRowCenter } from '@/styles/flex';
 import { toStringUsingDOMParser } from '@/shared/lib/toStringUsingDOMParser';
 import { convertDateString } from '@/shared/lib/convertDateString';
 import { CalendarDays, Eye } from 'lucide-react';
+import Link from 'next/link';
 
-const NewsList = ({ newsData }: { newsData: any }) => {
+const NewsPage = ({ newsData }: { newsData: any }) => {
   return (
-    <div className={`${flexRow} h-[180px] gap-4 my-4 px-4 py-4 bg-[#ECEEF2] cursor-pointer rounded-lg shadow-lg`}>
+    <Link href={`/media/wiznews/${newsData.artcSeq}`} className={`${flexRow} h-[180px] gap-4 my-4 px-4 py-4 bg-[#ECEEF2] cursor-pointer rounded-lg shadow-sm shadow-slateGray`}>
       <div className={`${flexRowCenter} w-1/4 h-full`}>
         <Image src={newsData.imgFilePath} alt={newsData.artcTitle} width="350" height="250" quality={75} />
       </div>
@@ -23,10 +24,9 @@ const NewsList = ({ newsData }: { newsData: any }) => {
           <span className={`${flexRowCenter} w-12 text-xs text-slateGray`}><Eye color="#717781"
                                                                            className="h-4" />{newsData.viewCnt}</span>
         </div>
-        <div className="border-t-[1px] border-slateGray" />
       </div>
-    </div>
+    </Link>
   );
 };
 
-export default NewsList;
+export default NewsPage;
